python/n1s_survey/requirements.txt
Django==3.1.1
djangorestframework
django-truncate
mysql
Pillow
